refactor(shared): hoist gen/enter twin infra modules into shared/
Moves the seven near-identical infrastructure modules (error, events,
util, logger, cache, middleware/validator, utils/api-docs) duplicated
between gen.pollinations.ai and enter.pollinations.ai into shared/ as
the single source of truth, and rewrites all imports in both services.

- error.ts: unified on the gen superset (errorCode override + 426
  status entry, both no-ops for enter); keeps enter's doc comments.
- events.ts: keeps gen's sendToTinybird + enter's live tier-event
  sender; drops gen's dead PolarEvent/flattenBalances/tier-event copy
  (zero callers since the #10466 gateway split, Polar removed #10625).
- util.ts: common helpers move to shared; gen keeps its service-only
  generateRandomId/safeRound/arrayBufferToBase64; enter's util.ts
  (a strict subset) is deleted.
- cache.ts: unified on the unknown[] generics; shared/utils/model-stats
  now imports cached directly, removing the CachedFn injection
  workaround and its extra parameter at both call sites.
- middleware/logger.ts stays per-service (gen redacts credential query
  params; semantic drift kept local).
