+++ enter.pollinations.ai/src/env.ts
@@ -1,9 +1,6 @@
+import type { ErrorVariables } from "@shared/error.ts";
 import type { RequestIdVariables } from "hono/request-id";
 import type { LoggerVariables } from "./middleware/logger.ts";
-
-export type ErrorVariables = {
-    error?: Error;
-};
 
 export type Env = {
     Bindings: CloudflareBindings;
