+++ enter.pollinations.ai/test/fixtures.ts
@@ -2,6 +2,7 @@ import { env, SELF } from "cloudflare:test";
 import type { Logger } from "@logtape/logtape";
 import { getLogger } from "@logtape/logtape";
 import { user as userTable } from "@shared/db/better-auth.ts";
+import { ensureConfigured } from "@shared/logger.ts";
 import {
     createFetchMock,
     teardownFetchMock,
@@ -11,7 +12,6 @@ import { createAuthClient } from "better-auth/client";
 import { adminClient, apiKeyClient } from "better-auth/client/plugins";
 import { drizzle } from "drizzle-orm/d1";
 import { test as base, expect } from "vitest";
-import { ensureConfigured } from "@/logger.ts";
 import { createMockGithub } from "./mocks/github.ts";
 import { createMockStripe } from "./mocks/stripe.ts";
