+++ gen.pollinations.ai/src/embeddings/openai.ts
@@ -1,6 +1,5 @@
+import { ensureUpstreamOk } from "@shared/error.ts";
 import type { Usage } from "@shared/registry/registry.ts";
-
-import { ensureUpstreamOk } from "@/error.ts";
 
 const OPENAI_EMBEDDINGS_ENDPOINT = "https://api.openai.com/v1/embeddings";
