+++ gen.pollinations.ai/src/env.ts
@@ -1,3 +1,4 @@
+import type { ErrorVariables } from "@shared/error.ts";
 import type { RequestIdVariables } from "hono/request-id";
 import type { AuthVariables } from "@/middleware/auth.ts";
 import type { BalanceVariables } from "@/middleware/balance.ts";
@@ -6,10 +7,6 @@ import type { FrontendKeyRateLimitVariables } from "@/middleware/rate-limit-dura
 import type { SafetyVariables } from "@/middleware/safety.ts";
 import type { ModelVariables } from "./middleware/model.ts";
 import type { TrackVariables } from "./middleware/track.ts";
-
-export type ErrorVariables = {
-    error?: Error;
-};
 
 export type Env = {
     Bindings: CloudflareBindings;
