+++ gen.pollinations.ai/src/middleware/track.ts
@@ -1,4 +1,5 @@
 import { getLogger } from "@logtape/logtape";
+import type { ApiKeyType } from "@shared/auth/api-key-creation.ts";
 import { AUTO_TOP_UP_THRESHOLD_POLLEN } from "@shared/billing/auto-top-up.ts";
 import { payerBucketToMeter } from "@shared/billing/balance.ts";
 import {
@@ -12,6 +13,13 @@ import {
     truncateIpToSubnet,
 } from "@shared/client-ip.ts";
 import { user as userTable } from "@shared/db/better-auth.ts";
+import type { ErrorVariables } from "@shared/error.ts";
+import {
+    getDefaultErrorMessage,
+    getErrorCode,
+    UpstreamError,
+} from "@shared/error.ts";
+import { sendToTinybird } from "@shared/events.ts";
 import { PUBLIC_URLS } from "@shared/public-urls.ts";
 import type { Usage } from "@shared/registry/registry.ts";
 import {
@@ -29,10 +37,9 @@ import {
     parseUsageHeaders,
 } from "@shared/registry/usage-headers.ts";
 import type {
-    ApiKeyType,
     EventType,
     GenerationEventContentFilterParams,
-    InsertGenerationEvent,
+    TinybirdEvent as InsertGenerationEvent,
 } from "@shared/schemas/generation-event.ts";
 import {
     contentFilterResultsToEventParams,
@@ -46,6 +53,7 @@ import {
     ContentFilterResultSchema,
     ContentFilterSeveritySchema,
 } from "@shared/schemas/openai.ts";
+import { getRoutePath, removeUnset } from "@shared/util.ts";
 import { eq } from "drizzle-orm";
 import type { DrizzleD1Database } from "drizzle-orm/d1";
 import { drizzle } from "drizzle-orm/d1";
@@ -54,18 +62,11 @@ import type { HonoRequest } from "hono";
 import { createMiddleware } from "hono/factory";
 import { z } from "zod";
 import { mergeContentFilterResults } from "@/content-filter.ts";
-import type { ErrorVariables } from "@/env.ts";
-import {
-    getDefaultErrorMessage,
-    getErrorCode,
-    UpstreamError,
-} from "@/error.ts";
-import { sendToTinybird } from "@/events.ts";
 import type { AuthVariables } from "@/middleware/auth.ts";
 import type { BalanceVariables } from "@/middleware/balance.ts";
 import type { LoggerVariables } from "@/middleware/logger.ts";
 import type { FrontendKeyRateLimitVariables } from "@/middleware/rate-limit-durable.ts";
-import { generateRandomId, getRoutePath, removeUnset } from "@/util.ts";
+import { generateRandomId } from "@/util.ts";
 
 type ModelVariables = {
     model: {
