+++ gen.pollinations.ai/test/error-observability.test.ts
@@ -2,12 +2,12 @@ import {
     createExecutionContext,
     waitOnExecutionContext,
 } from "cloudflare:test";
+import { handleError, UpstreamError } from "@shared/error.ts";
 import { Hono } from "hono";
 import { HTTPException } from "hono/http-exception";
 import { requestId } from "hono/request-id";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import type { Env } from "@/env.ts";
-import { handleError, UpstreamError } from "@/error.ts";
 import { logger } from "@/middleware/logger.ts";
 import { handleChatCompletionLocal } from "@/text/handler.ts";
